local_requirements.txt
fabric
apache-libcloud
